requirements.txt
fastapi
uvicorn
torch
transformers
pillow